update 視線取得の関数化におけるファイルの追加および変更 現状、カメラが立ち上がるがすぐ落ちるという問題が未解決
