+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "python.analysis.extraPaths": [
         "./util"
-    ]
+    ],
+    "git.ignoreLimitWarning": true
 }
