+++ get_gaze_vec.py
@@ -2,7 +2,7 @@ def get_gaze_vec():
     from ax_gaze_estimation import ax_gaze_estimation_functionalize
     import judging_out
 
-    gaze_vec = ax_gaze_estimation_functionalize.main("--video 0")
+    gaze_vec = ax_gaze_estimation_functionalize.main()
 
 
     # judging_out.judging_out(gaze_vec)
